Readme: Explain Authorization headers
Now that #151 is merged, this adds a few lines of explanation as to how Authorization headers are sent for those of us whose APIs don't accept the "Bearer" prefix, such as Amazon AppSync with Auth0. When I was getting going with this library I found there were a few gotchas here, so this makes the explanation about two config variables more explicit.