fix: server side renderings
In https://github.com/nuxt-community/apollo-module/pull/336 the server side rendering was broken.
The problem is that when assigning the apollo cache to the nuxt state,
in `beforeNuxtRender` the state is actually still empty.
It only works because a reference is assigned there, which is later populated.

In https://github.com/nuxt-community/apollo-module/pull/336  another problem was fixed: Before this fix everything which was populated in the cache before `beforeNuxtRender` was cleared.

To fix both issues I propose to remove the workaround introduced in https://github.com/nuxt-community/apollo-module/pull/274 for the case when the `cache` object is shared across requests.
Since the server is handling requests in parallel, its simply impossible to have a shared cache instance.
Users of `nuxtjs/apollo` need to make sure that the cache instance is not shared, there is no other way.

That's why I propose to delete this workaround code, to fix all issues at the same time.

Issue: https://github.com/nuxt-community/apollo-module/issues/352