fix(nuxt-bridge): add support for nuxt bridge
Nuxt Bridge exposes the Nuxt 2 context via nuxtApp.nuxt2Context.

However, when doing `app.apolloProvider`, the apolloProvider is not added to this context.

In order to get the apolloProvider onto this nuxt2Context, we must use inject().

However, it requires a different name than "apolloProvider" when injecting, or we get errors.